+++ R/ctStanFit.R
@@ -151,7 +151,7 @@ T0VARredundancies <- function(ctm) { #check for redundant T0VAR parameters (beca
 #' @examples
 #' \donttest{
 #' 
-#' #generate a ctStanModel relying heavily on defaults
+#' #generate a modern ctsem model relying heavily on defaults
 #' model<-ctModel(type='ct',
 #'   latentNames=c('eta1','eta2'),
 #'   manifestNames=c('Y1','Y2'),